+++ crates/cargo-msrv-reporter/src/event/fetch_index.rs
@@ -70,4 +70,18 @@ mod tests {
             vec![Event::unscoped(Message::FetchIndex(event)),]
         );
     }
+
+    #[cfg(feature = "rust-releases-offline-source")]
+    #[test]
+    fn reported_offline_unless_outdated_source() {
+        let reporter = TestReporterWrapper::default();
+        let event = FetchIndex::new(ReleaseSource::OfflineUnlessOutdated);
+
+        reporter.get().report_event(event.clone()).unwrap();
+
+        assert_eq!(
+            reporter.wait_for_events(),
+            vec![Event::unscoped(Message::FetchIndex(event)),]
+        );
+    }
 }